python/requirements/base.txt
# SHA1:03d3fd323028c0c8f13099a78a5e158e2c27477e
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
attrs==25.3.0
    # via glom
boltons==25.0.0
    # via
    #   face
    #   glom
boto3==1.38.39
    # via -r requirements/base.in
botocore==1.38.39
    # via
    #   boto3
    #   s3transfer
cachetools==5.5.2
    # via google-auth
certifi==2025.6.15
    # via requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via -r requirements/base.in
crcmod==1.7
    # via -r requirements/base.in
face==24.0.0
    # via glom
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via -r requirements/base.in
glom==24.11.0
    # via -r requirements/base.in
google-api-core[grpc]==2.25.1
    # via
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.40.3
    # via
    #   google-api-core
    #   google-cloud-bigquery-storage
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-bigquery-storage==2.32.0
    # via -r requirements/base.in
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==2.19.0
    # via -r requirements/base.in
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio==1.73.0
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.73.0
    # via google-api-core
idna==3.10
    # via requests
importlib-resources==6.5.2
    # via -r requirements/base.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-bigquery-storage
protobuf==6.31.1
    # via
    #   google-api-core
    #   google-cloud-bigquery-storage
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
py4j==0.10.9.7
    # via pyspark
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pygments==2.19.1
    # via rich
pyspark==3.5.4
    # via -r requirements/base.in
python-dateutil==2.9.0.post0
    # via botocore
requests==2.32.4
    # via
    #   google-api-core
    #   google-cloud-storage
rich==14.0.0
    # via -r requirements/base.in
rsa==4.9.1
    # via google-auth
s3transfer==0.13.0
    # via boto3
six==1.17.0
    # via
    #   python-dateutil
    #   thrift
smmap==5.0.2
    # via gitdb
sqlglot==26.29.0
    # via -r requirements/base.in
thrift==0.21.0
    # via -r requirements/base.in
urllib3==2.5.0
    # via
    #   botocore
    #   requests
